fix waiting state for showcasing teams
