+++ shared/profile/user/index.tsx
@@ -7,7 +7,7 @@ import Assertion from '../../tracker2/assertion/container'
 import Bio from '../../tracker2/bio/container'
 import Friend from './friend/container'
 import Measure from './measure'
-import Teams from './teams/container'
+import Teams from './teams'
 import chunk from 'lodash/chunk'
 import * as T from '../../constants/types'
 import type {RPCError} from '../../util/errors'
