+++ shared/profile/user/teams/index.tsx
@@ -1,25 +1,71 @@
+import * as C from '../../../constants'
+import * as Constants from '../../../constants/tracker2'
+import * as T from '../../../constants/types'
 import * as Kb from '../../../common-adapters'
 import * as React from 'react'
-import type * as T from '../../../constants/types'
 import OpenMeta from './openmeta'
-import TeamInfo from './teaminfo'
+import {default as TeamInfo, type Props as TIProps} from './teaminfo'
 
-export type Props = {
-  // lint totally confused
-  teamShowcase: ReadonlyArray<T.Tracker.TeamShowcase>
-  teamMeta: {
-    [K in string]: {
+type OwnProps = {username: string}
+
+const noTeams = new Array<T.Tracker.TeamShowcase>()
+
+const Container = (ownProps: OwnProps) => {
+  const d = C.useTrackerState(s => Constants.getDetails(s, ownProps.username))
+  const _isYou = C.useCurrentUserState(s => s.username === ownProps.username)
+  const _roles = C.useTeamsState(s => s.teamRoleMap.roles)
+  const _teamNameToID = C.useTeamsState(s => s.teamNameToID)
+  const _youAreInTeams = C.useTeamsState(s => s.teamnames.size > 0)
+  const teamShowcase = d.teamShowcase || noTeams
+  const navigateAppend = C.useRouterState(s => s.dispatch.navigateAppend)
+  const _onEdit = () => {
+    navigateAppend('profileShowcaseTeamOffer')
+  }
+  const joinTeam = C.useTeamsState(s => s.dispatch.joinTeam)
+  const showTeamByName = C.useTeamsState(s => s.dispatch.showTeamByName)
+  const onJoinTeam = joinTeam
+  const clearModals = C.useRouterState(s => s.dispatch.clearModals)
+  const onViewTeam = (teamname: string) => {
+    clearModals()
+    showTeamByName(teamname)
+  }
+  const onEdit = _isYou && _youAreInTeams ? _onEdit : undefined
+  const teamMeta = teamShowcase.reduce<{
+    [key: string]: {
       inTeam: boolean
       teamID: T.Teams.TeamID
     }
-  }
-  onJoinTeam: (teamname: string) => void
-  onViewTeam: (teamname: string) => void
-  onEdit?: () => void
+  }>((map, t) => {
+    const teamID = _teamNameToID.get(t.name) || T.Teams.noTeamID
+    map[t.name] = {
+      inTeam: !!((_roles.get(teamID)?.role || 'none') !== 'none'),
+      teamID,
+    }
+    return map
+  }, {})
+
+  return onEdit || teamShowcase.length > 0 ? (
+    <Kb.Box2 direction="vertical" gap="tiny" fullWidth={true} style={styles.showcases}>
+      <Kb.Box2 direction="horizontal" gap="tiny" fullWidth={true}>
+        <Kb.Text type="BodySmallSemibold">Teams</Kb.Text>
+        {!!onEdit && <Kb.Icon type="iconfont-edit" onClick={onEdit} />}
+      </Kb.Box2>
+      {!!onEdit && !teamShowcase.length && <ShowcaseTeamsOffer onEdit={onEdit} />}
+      {teamShowcase.map(t => (
+        <TeamShowcase
+          key={t.name}
+          {...t}
+          onJoinTeam={onJoinTeam}
+          onViewTeam={() => onViewTeam(t.name)}
+          inTeam={teamMeta[t.name]?.inTeam ?? false}
+        />
+      ))}
+    </Kb.Box2>
+  ) : null
 }
 
-// TODO types
-const TeamShowcase = (props: any) => {
+const TeamShowcase = (props: Omit<TIProps, 'visible' | 'onHidden'>) => {
+  const {name, isOpen} = props
   const makePopup = React.useCallback(
     (p: Kb.Popup2Parms) => {
       const {attachTo, toggleShowingPopup} = p
@@ -36,16 +82,15 @@ const TeamShowcase = (props: any) => {
           <Kb.Avatar size={32} teamname={props.name} isTeam={true} />
         </>
         <Kb.Text type="BodySemiboldLink" style={styles.link}>
-          {props.name}
+          {name}
         </Kb.Text>
-        <OpenMeta isOpen={props.isOpen} />
+        <OpenMeta isOpen={isOpen} />
       </Kb.Box2>
     </Kb.ClickableBox>
   )
 }
 
-// TODO types
-const ShowcaseTeamsOffer = (p: any) => (
+const ShowcaseTeamsOffer = (p: {onEdit: () => void}) => (
   <Kb.Box2 direction="horizontal" gap="tiny" fullWidth={true}>
     <Kb.ClickableBox onClick={p.onEdit}>
       <Kb.Box2 direction="horizontal" gap="tiny">
@@ -57,26 +102,6 @@ const ShowcaseTeamsOffer = (p: any) => (
     </Kb.ClickableBox>
   </Kb.Box2>
 )
-
-const Teams = (p: Props) =>
-  p.onEdit || p.teamShowcase.length > 0 ? (
-    <Kb.Box2 direction="vertical" gap="tiny" fullWidth={true} style={styles.showcases}>
-      <Kb.Box2 direction="horizontal" gap="tiny" fullWidth={true}>
-        <Kb.Text type="BodySmallSemibold">Teams</Kb.Text>
-        {!!p.onEdit && <Kb.Icon type="iconfont-edit" onClick={p.onEdit} />}
-      </Kb.Box2>
-      {!!p.onEdit && !p.teamShowcase.length && <ShowcaseTeamsOffer onEdit={p.onEdit} />}
-      {p.teamShowcase.map(t => (
-        <TeamShowcase
-          key={t.name}
-          {...t}
-          onJoinTeam={p.onJoinTeam}
-          onViewTeam={() => p.onViewTeam(t.name)}
-          inTeam={p.teamMeta[t.name]?.inTeam ?? false}
-        />
-      ))}
-    </Kb.Box2>
-  ) : null
 
 const styles = Kb.Styles.styleSheetCreate(
   () =>
@@ -97,4 +122,4 @@ const styles = Kb.Styles.styleSheetCreate(
     }) as const
 )
 
-export default Teams
+export default Container
